Software/requirements.txt
pyserial
texttable
